extract search function from useNote
